Update camera.h
fix on var type

+++ src/game/camera.h
@@ -713,7 +713,7 @@ extern u8 gRecentCutscene;
 // TODO: sort all of this extremely messy shit out after the split
 
 void set_camera_shake_from_hit(s16 shake);
-void set_environmental_camera_shake(s16 shake);
+void set_environmental_camera_shake(s32 shake);
 void set_camera_shake_from_point(s16 shake, f32 posX, f32 posY, f32 posZ);
 void move_mario_head_c_up(UNUSED struct Camera *c);
 void transition_next_state(UNUSED struct Camera *c, s16 frames);
